connectivity_check/requirements.txt
# always get latest
harness-featureflags
fastapi==0.90.1
uvicorn[standard]